Reworked getting user email
Removed the `get_email_form_id_token` method from AuthHelper module.
This method relied on parsing the ID token and assuming that the
`preferred_username` claim was equal to the user's email address. In
many cases this is true, but it isn't guaranteed.

Replaced this with the `get_user_email` method, which calls the Outlook
API, doing a GET on `/api/v2.0/Me' to get the user's information,
including the email address.

This also removes the necessity of including the `profile` scope in the
authorization request.

Resolves #6